\name{arrows.default}
\alias{arrows.default}
\title{Draw default arrows}
\usage{
arrows.default(coordinates, ...)
}
\arguments{
  \item{coordinates}{length-4 vector giving x0, x1, y0, y1
  from arrows (in that order)}

  \item{...}{Further commands passed to arrows}
}
\value{
draws an arrow on the current plot
}
\description{
Simple internal function to make basic arrows appear in a
standard manner
}

