\name{binary.entropy}
\alias{binary.entropy}
\title{Binary entropy of one or more proportions}
\usage{
binary.entropy(x)
}
\arguments{
  \item{x}{value (or vector of values) between zero and one}
}
\value{
vector of values representing the Shannon information of each entry.
}
\description{
Uses binary entropy formula - useful in conjunciton with prop.occupied.
}

