\name{draw.line.key}
\title{Draw a key to line colours/widths}
\usage{
draw.line.key(input, labels)
}
\arguments{
  \item{input}{an object (invisibly) returned by plot.spaa}
  \item{labels}{an optional list of labels to override the defaults}
}
\value{
Draws a (basic) key to line colours and widths
}
\seealso{
draw.point.key draw.spp.key
}

