\name{draw.object.key}
\alias{draw.object.key}
\title{Sets key behaviour by plot.pairs()}
\usage{
draw.object.key(draw.frequencies)
}
\arguments{
  \item{draw.frequencies}{Should species occupancy rates be drawn as point sizes/colours? Defaults to TRUE}
}
\value{
set window behaviour using split.screen()
}
\description{
Internal function to set arrangement of 'object' keys (i.e. line and point keys)
}

