\name{draw.point.key}
\alias{draw.point.key}
\title{Draw a key to point colours/widths}
\usage{
draw.point.key(input)
}
\arguments{
  \item{input}{an object (invisibly) returned by plot.spaa}
}
\value{
Draws a (basic) key to point colours and widths
}
\seealso{
	\code{\link{draw.line.key}}, \code{\link{draw.spp.key}}
}
