\name{draw.spp.key}
\alias{draw.spp.key}
\title{Draw a key to species names}
\usage{
draw.object.key(points, draw.frequencies)
}
\arguments{
  \item{points}{a dataframe given by or.points}
  \item{draw.frequencies}{Should species occupancy rates be drawn as point sizes/colours? Defaults to TRUE}
}
\value{
set window behaviour using split.screen()
}
\description{
Internal function to draw a key to species labels (names).
}

