\name{odds.ratio.test}
\alias{odds.ratio.test}
\title{Odds Ratio Test}
\usage{
odds.ratio.test(dataset)
}
\arguments{
  \item{dataset}{a two-column matrix or data.frame
  containing presence/absence data for two species}
}
\value{
Gives either an error message, or invisibly returns
'dataset' (possibly corrected by 'make.binary')
}
\description{
Odds Ratio Test
}

