\name{offset.line}
\alias{offset.line}
\title{Line perpendicular offset}
\usage{
offset.line(coordinates, offset)
}
\arguments{
  \item{coordinates}{length-4 vector of form c(x0, x1, y0,
  y1)}

  \item{offset}{how far to shift lines? default = 0.2}
}
\description{
Function to perpendicularly offset lines, for cases where
there are arrows in both directions. Not yet implemented.
}

