\name{or.asymmetric}
\alias{or.asymmetric}
\title{Asymmetric odds ratio calculation using a contingency table}
\usage{
or.asymmetric(dataset)
}
\arguments{
  \item{dataset}{Two column data.frame, with presence/absence of two species}
}
\value{
value of the odds ratio between those species
}
\description{
Calculate asymmetric odds ratio from a contingency table
}

