\name{or.contingency}
\alias{or.contingency}
\title{Odds ratio calculation using a contingency table}
\usage{
or.contingency(dataset)
}
\arguments{
  \item{dataset}{Two column data.frame, with
  presence/absence for two species}
}
\value{
value of the odds ratio between those species
}
\description{
Calculate asymmetric odds ratio from a contingency table
}

