\name{or.symmetric}
\alias{or.symmetric}
\title{Symmetric odds ratio calculation using a contingency table}
\usage{
or.symmetric(dataset)
}
\arguments{
  \item{dataset}{Two column data.frame, with presence/absence of two species}
}
\value{
value of the odds ratio between those species
}
\description{
Calculate symmetric odds ratio from a contingency table
}

