\name{binary.entropy}
\alias{binary.entropy}
\title{Binary entropy of proportions data}
\usage{
binary.entropy(x)
}
\arguments{
  \item{x}{Value (or vector of values) between zero and one}
}
\value{
Vector of values representing the Shannon information of each entry.
}
\description{
Uses binary entropy formula - useful in conjunction with prop.occupied.
}

