\name{draw.spp.key}
\title{Draw a key to species}
\usage{
draw.spp.key(input, labels)
}
\arguments{
  \item{input}{an object (invisibly) returned by plot.spaa}
  \item{labels}{an optional list of labels to override the defaults}
}
\value{
Draws a (basic) key to the species in a larger plot.
}
\seealso{
draw.line.key draw.point.key
}