\name{prop.occupied}
\alias{prop.occupied}
\title{Propotion of occupied sites}
\usage{
prop.occupied(dataset)
}
\arguments{
  \item{dataset}{site by species matrix (stored as either a data.frame or matrix) containing presence/absence data}
}
\value{
vector of proportions
}
\description{
Calculate the proportion of ones in each column of a data.frame or matrix
}

